src/requirement.txt
mysql-connector-python
bcrypt # for password hashing